app/requirements.txt
acres==0.5.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
configobj==5.0.9
configparser==7.2.0
cryptography==45.0.5
etelemetry==0.3.1
Faker==37.4.0
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.2
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
pathlib==1.0.1
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prov==2.1.1
psycopg2-binary==2.9.10
puremagic==1.29
pycparser==2.22
pydot==4.0.1
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
sqlite-vec==0.1.7a2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
